requirements.txt
jinja2~=3.1.6
pydantic~=2.11.7
python-dotenv~=1.1.1
python-multipart~=0.0.20
requests~=2.32.4
starlette~=0.47.2
uvicorn~=0.35.0
